Documentation set added. Downloading runs only if file inexists

+++ config.php
@@ -4,17 +4,20 @@ return [
     'server'           => "127.0.0.1:4444",
     'downloadFunction' => function ($link, $file)
     {
-        $curl = curl_init();
-        curl_setopt($curl, CURLOPT_SSL_VERIFYPEER, false);
-        curl_setopt($curl, CURLOPT_HEADER, false);
-        curl_setopt($curl, CURLOPT_FOLLOWLOCATION, true);
-        curl_setopt($curl, CURLOPT_URL, $link);
-        curl_setopt($curl, CURLOPT_REFERER, $link);
-        curl_setopt($curl, CURLOPT_RETURNTRANSFER, true);
-        $content = curl_exec($curl);
-        curl_close($curl);
-        
-        file_put_contents($file, $content, LOCK_EX);
+        if (!file_exists($file))
+        {
+            $curl = curl_init();
+            curl_setopt($curl, CURLOPT_SSL_VERIFYPEER, false);
+            curl_setopt($curl, CURLOPT_HEADER, false);
+            curl_setopt($curl, CURLOPT_FOLLOWLOCATION, true);
+            curl_setopt($curl, CURLOPT_URL, $link);
+            curl_setopt($curl, CURLOPT_REFERER, $link);
+            curl_setopt($curl, CURLOPT_RETURNTRANSFER, true);
+            $content = curl_exec($curl);
+            curl_close($curl);
+
+            file_put_contents($file, $content, LOCK_EX);
+        }
     },
     'commandExecutor'  => function ($commands)
     {
@@ -56,7 +59,7 @@ return [
             }
             else
             {
-
+                
                 $iterator = new RecursiveDirectoryIterator($path, RecursiveDirectoryIterator::SKIP_DOTS);
                 $files = new RecursiveIteratorIterator(
                     $iterator, RecursiveIteratorIterator::CHILD_FIRST
@@ -76,4 +79,13 @@ return [
             }
         }
     },
+    'clean'            => [
+        'composer.phar',
+        'composer.lock',
+        'codecept.phar',
+        'composer-installer.php',
+        'vendor',
+        'tests' . DIRECTORY_SEPARATOR . '_output' . DIRECTORY_SEPARATOR . 'c3tmp',
+        'c3.php',
+    ],
 ];
